Screen simulating, command receiving
Implemented the receiving of data.

Codes cleanups all around.

Command clarifications.

The arduino can now simulate the rollout / rollup of a screen by using
the LEDs on the breadboard.
Green	=>	open
Red	=>	closed
Yellow	=>	opening/closing

While opening or closing, the yellow light will flash until the screen
is opened / closed. The LED of the screen's previous state will stay on.

The Arduino is now also able to receive commands. It keeps checking to
see if one has been sent, and if one has, it will execute it. If the
command is not known, it will do nothing.

0x53	=>	open screen
0x52	=>	close screen